refactor(manager/download): inline ParallelStrategy into MultiPartDownloader (#187)
Removes xearthlayer/src/manager/download/strategy.rs entirely. The
ParallelStrategy struct existed solely to back the now-deleted
DownloadStrategy trait; with one concrete implementation and no other
callers, the indirection was load-bearing nostalgia rather than
architecture.

Changes:
- Move execute() body into MultiPartDownloader as a private
  run_downloads() method. Reads parallel_downloads and timeout
  directly from self instead of from a duplicate ParallelStrategy
  field.
- Move MAX_RETRIES and BASE_RETRY_DELAY constants into orchestrator.rs.
- Drop the three ParallelStrategy struct-construction tests; the
  equivalent MultiPartDownloader tests already cover concurrency,
  timeout, and minimum-1 clamping behaviour.
- Update download module doc and architecture diagram.

Net delta: -75 lines, -1 file.
