+++ xearthlayer/src/manager/download/orchestrator.rs
@@ -1,28 +1,39 @@
 //! Multi-part download orchestrator.
 //!
-//! This module provides the high-level orchestration for downloading
-//! multi-part packages, using configurable download strategies.
+//! Coordinates downloading all parts of a package using a
+//! semaphore-bounded sliding window. With `parallel_downloads = 1`,
+//! the semaphore reduces this to one active download at a time while
+//! keeping the same retry, progress, and (in subsequent commits in
+//! this branch) cancellation semantics.
 
 use std::sync::Arc;
 use std::thread;
 use std::time::Duration;
 
 use super::http::HttpDownloader;
-use super::progress::DownloadProgressCallback;
+use super::progress::{DownloadProgressCallback, ProgressCounters, ProgressReporter};
+use super::semaphore::CountingSemaphore;
 use super::state::DownloadState;
-use super::strategy::ParallelStrategy;
 use crate::manager::error::{ManagerError, ManagerResult};
-use crate::manager::traits::PackageDownloader;
+use crate::manager::traits::{PackageDownloader, ProgressCallback};
+
+/// Maximum number of attempts per part.
+///
+/// One initial attempt plus two retries.
+const MAX_RETRIES: u8 = 3;
+
+/// Base delay for exponential backoff (2 seconds).
+const BASE_RETRY_DELAY: Duration = Duration::from_secs(2);
 
 /// Multi-part download orchestrator.
 ///
-/// Coordinates the download of all parts of a package using a configurable
-/// strategy (sequential or parallel).
+/// Coordinates the download of all parts of a package with a
+/// semaphore-bounded sliding window of concurrent workers.
 #[derive(Debug)]
 pub struct MultiPartDownloader {
     /// HTTP downloader for individual files.
     downloader: HttpDownloader,
-    /// Number of parallel downloads (determines strategy).
+    /// Number of parallel downloads (minimum 1).
     parallel_downloads: usize,
 }
 
@@ -93,10 +104,9 @@ impl MultiPartDownloader {
 
     /// Download all parts of a package.
     ///
-    /// Always uses [`ParallelStrategy`]; for sequential behaviour, set
-    /// `parallel_downloads = 1` and the strategy's internal semaphore
-    /// gates work to one at a time. Returns `Ok(())` on success, or an
-    /// error if any part failed to download.
+    /// Returns `Ok(())` if every part completed successfully.
+    /// Returns [`ManagerError::DownloadFailed`] when one or more parts
+    /// failed to download after [`MAX_RETRIES`] attempts.
     pub fn download_all(
         &self,
         state: &mut DownloadState,
@@ -104,8 +114,7 @@ impl MultiPartDownloader {
     ) -> ManagerResult<()> {
         let on_progress = on_progress.map(Arc::new);
 
-        let strategy = ParallelStrategy::new(self.parallel_downloads, self.downloader.timeout);
-        strategy.execute(state, on_progress)?;
+        self.run_downloads(state, on_progress)?;
 
         // Check for failures
         if state.has_failures() {
@@ -114,6 +123,159 @@ impl MultiPartDownloader {
                 reason: format!("Parts {:?} failed to download", state.failed),
             });
         }
+
+        Ok(())
+    }
+
+    /// Spawn one worker per part, gate concurrency via a counting
+    /// semaphore, and collect results into `state` once all workers
+    /// terminate.
+    ///
+    /// Returns `Ok(())` for any normal completion (success, partial
+    /// failure, or all-failed). The caller maps `state.failed` into
+    /// an outward-facing error.
+    fn run_downloads(
+        &self,
+        state: &mut DownloadState,
+        on_progress: Option<Arc<DownloadProgressCallback>>,
+    ) -> ManagerResult<()> {
+        let total_parts = state.total_parts;
+
+        // Build filenames from destination paths
+        let filenames: Vec<String> = state
+            .destinations
+            .iter()
+            .map(|d| {
+                d.file_name()
+                    .map(|f| f.to_string_lossy().to_string())
+                    .unwrap_or_else(|| format!("part_{}", 0))
+            })
+            .collect();
+
+        // Create shared progress counters with extended metadata
+        let counters = Arc::new(ProgressCounters::new_extended(
+            total_parts,
+            filenames,
+            state.part_sizes.clone(),
+        ));
+
+        // Start progress reporter if callback provided
+        let _reporter = on_progress
+            .as_ref()
+            .map(|cb| ProgressReporter::start_detailed(Arc::clone(&counters), Arc::clone(cb)));
+
+        // Create semaphore for concurrency limiting
+        let semaphore = Arc::new(CountingSemaphore::new(self.parallel_downloads));
+
+        // Spawn all threads (semaphore gates actual concurrency)
+        let mut handles = Vec::with_capacity(total_parts);
+
+        for i in 0..total_parts {
+            let url = state.urls[i].clone();
+            let checksum = state.checksums[i].clone();
+            let dest = state.destinations[i].clone();
+            let counters = Arc::clone(&counters);
+            let semaphore = Arc::clone(&semaphore);
+            let timeout = self.downloader.timeout;
+
+            let handle = thread::spawn(move || {
+                // Acquire permit — blocks until a slot is available
+                let _permit = semaphore.acquire();
+
+                // Mark as downloading
+                counters.set_part_state(i, 1); // Downloading
+                counters.set_part_attempt(i, 1);
+
+                let mut last_error = String::new();
+
+                for attempt in 1..=MAX_RETRIES {
+                    if attempt > 1 {
+                        // Mark as retrying
+                        counters.set_part_state(i, 4); // Retrying
+                        counters.set_part_attempt(i, attempt);
+
+                        // Reset byte counter for retry
+                        counters.update_part(i, 0);
+
+                        // Exponential backoff: 2s, 4s, 8s
+                        let delay = BASE_RETRY_DELAY * 2u32.pow(attempt as u32 - 2);
+                        thread::sleep(delay);
+
+                        // Back to downloading
+                        counters.set_part_state(i, 1); // Downloading
+                    }
+
+                    let downloader = HttpDownloader::with_timeout(timeout);
+
+                    // Create per-download progress callback
+                    let counters_clone = Arc::clone(&counters);
+                    let part_index = i;
+                    let progress_cb: ProgressCallback =
+                        Box::new(move |downloaded: u64, _total: u64| {
+                            counters_clone.update_part(part_index, downloaded);
+                        });
+
+                    let result = downloader.download_with_progress(
+                        &url,
+                        &dest,
+                        Some(checksum.as_str()),
+                        progress_cb,
+                    );
+
+                    match result {
+                        Ok(bytes) => {
+                            counters.mark_completed(i, bytes);
+                            counters.set_part_state(i, 2); // Done
+                            return Some(bytes);
+                        }
+                        Err(e) => {
+                            last_error = e.to_string();
+                            // Remove partial file for clean retry
+                            std::fs::remove_file(&dest).ok();
+                        }
+                    }
+                }
+
+                // All retries exhausted
+                counters.set_part_state(i, 3); // Failed
+                counters.set_part_attempt(i, MAX_RETRIES);
+                counters.set_part_error(i, last_error);
+                None
+            });
+
+            handles.push((i, handle));
+        }
+
+        // Collect results
+        let mut failed_parts = Vec::new();
+        let mut total_bytes = 0u64;
+        let mut completed = 0usize;
+
+        for (index, handle) in handles {
+            match handle.join() {
+                Ok(Some(bytes)) => {
+                    total_bytes += bytes;
+                    completed += 1;
+                }
+                Ok(None) => {
+                    failed_parts.push(index);
+                }
+                Err(_) => {
+                    // Thread panicked
+                    failed_parts.push(index);
+                }
+            }
+        }
+
+        // Signal done and let reporter emit final snapshot
+        counters.signal_done();
+
+        // Update state with results
+        state.downloaded_parts = completed;
+        state.bytes_downloaded = total_bytes;
+        state.failed = failed_parts;
+
+        // Reporter will be dropped here, which stops it cleanly
 
         Ok(())
     }
